Phase C-1 cascade: move FrameInfo to majit-backend::resume_value
resume.py:250 jitcode_index + pc + slot_map structure carrying the
guard-time frame snapshot.  Depends only on FrameSlotSource (alias
for ResumeValueSource, already moved to majit-backend::resume_value
in 3171e412bb), so the move is a clean next hop in the ResumeData
migration chain.

FrameSlotSource type alias also moves to majit-backend; majit-metainterp
re-exports both for caller compatibility.
